Actualizaciones 1,3,6

+++ lib/screens/favourite_screen.dart
@@ -153,14 +153,14 @@ class _FavouriteScreenState extends State<FavouriteScreen> {
                             context.read<HomeCubit>().addProductData(
                                   item,
                                   data: data.copyWith(
-                                      quantity: num.parse(value).toInt()),
+                                      quantity: num.parse(value).toDouble()),
                                 );
                             context
                                 .read<EditProductBloc>()
                                 .add(EditProductAddProduct(product: item));
                           } else {
                             final updated = data.copyWith(
-                                quantity: num.parse(value).toInt());
+                                quantity: num.parse(value).toDouble());
                             context
                                 .read<HomeCubit>()
                                 .updateProductData(updated);
